+++ lib/reach/cli/analyses/smell/check.ex
@@ -1,0 +1,5 @@
+defmodule Reach.CLI.Analyses.Smell.Check do
+  @moduledoc false
+
+  @callback run(Reach.Project.t()) :: [map()]
+end